backend/requirements.txt
blinker==1.9.0
cachetools==6.2.0
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
dnspython==2.8.0
Flask==3.1.2
Flask-PyMongo==3.0.1
google-auth==2.41.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pymongo==4.15.2
python-dotenv==1.1.1
requests==2.32.5
rsa==4.9.1
urllib3==2.5.0
Werkzeug==3.1.3
